clippy lint

+++ src/algorithm/media.rs
@@ -434,7 +434,7 @@ pub fn value_to_image(value: &Value) -> Result<DynamicImage, String> {
 fn complex_color(c: Complex) -> [f64; 3] {
     let h = c.arg();
     let mag = c.abs();
-    let s = (0.3 + 0.7 * (-mag / 10.0).exp()) / (0.3 + 0.7 * (-0.1 as f64).exp());
+    let s = (0.3 + 0.7 * (-mag / 10.0).exp()) / (0.3 + 0.7 * (-0.1_f64).exp());
     let v = (1.0 - (-PI * mag).exp()) / (1.0 - (-PI).exp());
     hsv_to_rgb(h, s.min(1.0), v.min(1.0))
 }
